recycle: refact controller mock
