+++ src/modules/record/test/mocks/common.mock.ts
@@ -1,32 +1,32 @@
-import {
-  AllProfessionalRecords,
-  NewRecord,
-  ProfessionalRecord,
-} from '../../interfaces/record.interface';
-import {
-  mockPrismaNewRecord,
-  mockRepositoryRecordResponse,
-} from './repository.mock';
-import { mockEncryptedRecord, mockUserAppointment } from './service.mock';
+// import {
+//   AllProfessionalRecords,
+//   NewRecord,
+//   ProfessionalRecord,
+// } from '../../interfaces/record.interface';
+// import {
+//   mockPrismaNewRecord,
+//   mockRepositoryRecordResponse,
+// } from './repository.mock';
+// import { mockEncryptedRecord, mockUserAppointment } from './service.mock';
 
-export const mockNewRecord: NewRecord = {
-  recordId: mockRepositoryRecordResponse.id,
-  clientName: mockUserAppointment.clientName,
-  scheduledDate: mockUserAppointment.appointmentDate,
-  appointmentTime: mockUserAppointment.appointmentTime,
-  createdAt: mockRepositoryRecordResponse.createdAt,
-};
+// export const mockNewRecord: NewRecord = {
+//   recordId: mockRepositoryRecordResponse.id,
+//   clientName: mockUserAppointment.clientName,
+//   scheduledDate: mockUserAppointment.appointmentDate,
+//   appointmentTime: mockUserAppointment.appointmentTime,
+//   createdAt: mockRepositoryRecordResponse.createdAt,
+// };
 
-export const mockProfessionalRecord: ProfessionalRecord = {
-  recordId: mockRepositoryRecordResponse.id,
-  clientName: mockUserAppointment.clientName,
-  scheduledDate: mockUserAppointment.appointmentDate,
-  appointmentTime: mockUserAppointment.appointmentTime,
-  record: mockEncryptedRecord.record,
-  createdAt: mockRepositoryRecordResponse.createdAt,
-  updatedAt: mockPrismaNewRecord.updated_at,
-};
+// export const mockProfessionalRecord: ProfessionalRecord = {
+//   recordId: mockRepositoryRecordResponse.id,
+//   clientName: mockUserAppointment.clientName,
+//   scheduledDate: mockUserAppointment.appointmentDate,
+//   appointmentTime: mockUserAppointment.appointmentTime,
+//   record: mockEncryptedRecord.record,
+//   createdAt: mockRepositoryRecordResponse.createdAt,
+//   updatedAt: mockPrismaNewRecord.updated_at,
+// };
 
-export const mockAllProfessionalRecords: AllProfessionalRecords = {
-  records: [mockProfessionalRecord],
-};
+// export const mockAllProfessionalRecords: AllProfessionalRecords = {
+//   records: [mockProfessionalRecord],
+// };
